Moved common multitask code into methods. Changed multitask test to scala suite convention
